Refactor API routers and fix broad exception handling (Issues #3 & #4)
- Fix Issue #4: Replace broad exception handlers with specific exception types
  * colors.py: Catch ValueError/TypeError for float conversion
  * colors.py: Catch json.JSONDecodeError for JSON parsing
  * colors.py: Catch requests.RequestException for HTTP errors
  * spacing.py: Catch anthropic.APIError and requests.RequestException
  * projects.py, snapshots.py: Catch json.JSONDecodeError
  * validators.py: Catch binascii.Error and OSError/IOError

- Fix Issue #3: Extract business logic from colors.py router
  * Move get_extractor() to colors_service.py
  * Move parse_metadata() to colors_service.py
  * Move find_accent_hex() to colors_service.py
  * Move post_process_colors() to colors_service.py
  * Move default_shadow_tokens() to colors_service.py
  * Move add_role_tokens() to colors_service.py
  * Keep router-specific helpers in colors.py
  * Update all function calls to use service versions

Results:
- colors.py: 965 → 798 lines (167 lines removed, 17% reduction)
- colors_service.py: 185 → 361 lines (business logic consolidated)
- All 122 API tests passing ✅
